Extract out renaming logic into FileSystemViewFileName
